Adds a separate crate, packer_derive, which packer imports.
This allows lazy_static to be referenced.
