+++ README.md
@@ -1,8 +1,7 @@
 Packer
 ======
 
-[![](https://api.travis-ci.org/iptq/packer.svg?branch=master)](https://travis-ci.org/iptq/packer)
-[![dependency status](https://deps.rs/repo/github/iptq/packer/status.svg)](https://deps.rs/repo/github/iptq/packer)
+[![](https://api.travis-ci.org/iptq/packer.svg?branch=master)](https://travis-ci.org/iptq/packer) [![dependency status](https://deps.rs/repo/github/iptq/packer/status.svg)](https://deps.rs/repo/github/iptq/packer)
 
 **NOTE** This project is based on [the original rust-embed](https://github.com/pyros2097/rust-embed). I made enough modifications to it that I decided to just publish it in a separate repository.
 
@@ -14,7 +13,7 @@ Include the crate in your `Cargo.toml`:
 
 ```toml
 [dependencies]
-packer = "0.1"
+packer = "0.2"
 ```
 
 and in your `lib.rs` or `main.rs`:
@@ -26,8 +25,7 @@ extern crate packer;
 
 ### Step 2: Derive
 
-Start deriving `Packer` from your structs. You need to provide a `folder` attribute to indicate the folder 
-from which it should be pulling. Paths are relative to the crate root.
+Start deriving `Packer` from your structs. You need to provide a `folder` attribute to indicate the folder from which it should be pulling. Paths are relative to the crate root.
 
 ```rs
 #[derive(Packer)]
@@ -40,13 +38,13 @@ struct Assets;
 You can now access any file using the `get` function:
 
 ```rs
-Assets::get("kermit.jpg"); // this is a Option<Vec<u8>>
+let data: Option<&'static [u8]> = Assets::get("kermit.jpg");
 ```
 
 You may also choose to list all the files that have been stored.
 
 ```rs
-Assets::list(); // this is a ::std::vec::IntoIter<&'static str>
+let files /*: impl Iterator<Item = &'static str>*/ = Assets::list();
 ```
 
 When you build in dev mode, it will fetch off your filesystem as usual, but when you build with `--release`, it will pack the assets into your binary!
@@ -54,7 +52,7 @@ When you build in dev mode, it will fetch off your filesystem as usual, but when
 Future Work
 -----------
 
-- Possibly add options for excluding files?
+-	Possibly add options for excluding files?
 
 Contact
 -------
